requirements.txt
tensorflow-gpu>=2.0.1
pandas
sklearn